+++ js/modals.js
@@ -2,7 +2,6 @@ const ALERT_SHOW_TIME = 5000;
 
 const errorTemplate = document.querySelector('#error').content.querySelector('.error');
 const successTemplate = document.querySelector('#success').content.querySelector('.success');
-const imgUploadOverlay = document.querySelector('.img-upload__overlay');
 
 const successFragment = document.createDocumentFragment();
 const errorFragment = document.createDocumentFragment();
@@ -62,31 +61,3 @@ export const showAlertMessage = (message) => {
     alertContainer.remove();
   }, ALERT_SHOW_TIME);
 };
-
-export const showUploadPopup = (cb, isReadyForClose, cbClose) => {
-  imgUploadOverlay.classList.remove('hidden');
-  document.body.classList.add('modal-open');
-  document.addEventListener('keydown', (evt) => keyDownHandler(evt, isReadyForClose, cbClose));
-  if (cb) {
-    cb();
-  }
-};
-
-export const closeUploadPopup = (cb, isReadyForClose) => {
-  imgUploadOverlay.classList.add('hidden');
-  document.body.classList.remove('modal-open');
-  document.removeEventListener('keydown', (evt) => keyDownHandler(evt, isReadyForClose, cb));
-  if (cb) {
-    cb();
-  }
-};
-
-//Функция объявлена декларативно, чтобы могла быть вызвана раньше, чем она объявлена
-function keyDownHandler (evt, isReadyForClose, cbClose) {
-  if (evt.key === 'Escape' && isReadyForClose) {
-    evt.preventDefault();
-    if (cbClose) {
-      closeUploadPopup(cbClose, isReadyForClose);
-    }
-  }
-}
